testing BDDNetworkGraph

+++ projects/batfish/src/main/java/org/batfish/symbolic/bdd/BDDInteger.java
@@ -74,7 +74,7 @@ public class BDDInteger {
     for (int i = 0; i < _bitvec.length; i++) {
       BDD bitBDD = _bitvec[_bitvec.length - i - 1];
       if (!satAssignment.and(bitBDD).isZero()) {
-        value |= 1 << i;
+        value |= ((long) 1) << i;
       }
     }
     return value;
